chore(internal): move githelpers to internal/git
