+++ internal/sidekick/rust_release/preflight.go
@@ -20,7 +20,7 @@ import (
 	"log/slog"
 
 	"github.com/googleapis/librarian/internal/command"
-	"github.com/googleapis/librarian/internal/librarian/githelpers"
+	"github.com/googleapis/librarian/internal/git"
 	"github.com/googleapis/librarian/internal/sidekick/config"
 )
 
@@ -46,10 +46,10 @@ func CargoPreFlight(ctx context.Context, config *config.Release) error {
 // PreFlight() verifies all the necessary  tools are installed.
 func PreFlight(ctx context.Context, config *config.Release) error {
 	gitExe := gitExe(config)
-	if err := githelpers.GitVersion(ctx, gitExe); err != nil {
+	if err := git.GitVersion(ctx, gitExe); err != nil {
 		return err
 	}
-	if err := githelpers.GitRemoteURL(ctx, gitExe, config.Remote); err != nil {
+	if err := git.GitRemoteURL(ctx, gitExe, config.Remote); err != nil {
 		return err
 	}
 	if err := CargoPreFlight(ctx, config); err != nil {
